range_chmin_range_min のデータの型を usize から i64 にした

+++ src/mylib/segtree_lib/lazy_segtree/range_chmin_range_min.rs
@@ -10,10 +10,10 @@ pub mod range_chmin_range_min {
 
     pub struct RangeChminRangeMin(Infallible);
     impl MapMonoid for RangeChminRangeMin {
-        type M = Min<usize>;
-        type F = usize;
+        type M = Min<i64>;
+        type F = i64;
         fn identity_map() -> Self::F {
-            usize::MAX
+            i64::MAX
         }
         fn mapping(
             f: &Self::F,
